refactor: inline dashboard client logic
